% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialConversion.R
\name{AddGenericSpatialData}
\alias{AddGenericSpatialData}
\title{Add generic spatial data to Seurat object}
\usage{
AddGenericSpatialData(
  seurat_obj,
  coords,
  assay_name = "Spatial",
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{coords}{Spatial coordinates}

\item{assay_name}{Assay name}

\item{verbose}{Print messages}
}
\value{
Modified Seurat object
}
\description{
Add generic spatial data to Seurat object
}
\keyword{internal}
