% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialConversion.R
\name{AddVisiumSpatialData}
\alias{AddVisiumSpatialData}
\title{Add Visium spatial data to Seurat object}
\usage{
AddVisiumSpatialData(
  seurat_obj,
  coords,
  visium_data,
  h5ad,
  assay_name = "Spatial",
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{coords}{Spatial coordinates}

\item{visium_data}{Visium-specific metadata}

\item{assay_name}{Assay name}

\item{verbose}{Print messages}
}
\value{
Modified Seurat object
}
\description{
Add Visium spatial data to Seurat object
}
\keyword{internal}
