% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{as.factor.H5D}
\alias{as.factor.H5D}
\title{Convert an HDF5 dataset to a factor}
\usage{
as.factor.H5D(x)
}
\arguments{
\item{x}{An HDF5 dataset}
}
\value{
A factor
}
\description{
Convert an HDF5 dataset to a factor
}
\keyword{internal}
