% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{as.sparse.H5D}
\alias{as.sparse.H5D}
\alias{as.sparse.H5Group}
\title{Convert an HDF5 dataset to a sparse matrix}
\usage{
\method{as.sparse}{H5D}(x)

\method{as.sparse}{H5Group}(x)
}
\arguments{
\item{x}{An HDF5 dataset or group}
}
\value{
A sparse matrix
}
\description{
Convert an HDF5 dataset to a sparse matrix
}
