% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{CompoundToGroup}
\alias{CompoundToGroup}
\title{Convert compound dataset to group}
\usage{
CompoundToGroup(src, dst, dname, order, index = NULL, name_map_fn = NULL)
}
\arguments{
\item{src}{Source compound dataset}

\item{dst}{Destination file}

\item{dname}{Destination name}

\item{order}{Order attribute name}

\item{index}{Index for subsetting}
}
\description{
Convert compound dataset to group
}
\keyword{internal}
