% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ConvertBPCellsMatrix}
\alias{ConvertBPCellsMatrix}
\title{Convert BPCells matrix objects to dgCMatrix}
\usage{
ConvertBPCellsMatrix(mat, verbose = FALSE)
}
\arguments{
\item{mat}{Matrix object (potentially BPCells IterableMatrix or RenameDims)}

\item{verbose}{Show conversion message}
}
\value{
dgCMatrix or original matrix if not BPCells
}
\description{
Convert BPCells matrix objects to dgCMatrix
}
\keyword{internal}
