% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFOV.R
\name{ConvertFOVToH5AD}
\alias{ConvertFOVToH5AD}
\title{Convert FOV object to h5ad-compatible structure}
\usage{
ConvertFOVToH5AD(fov_obj, verbose = TRUE)
}
\arguments{
\item{fov_obj}{FOV object from Seurat}

\item{verbose}{Show progress messages}
}
\value{
List with spatial data in h5ad-compatible format
}
\description{
Convert FOV object to h5ad-compatible structure
}
\keyword{internal}
