% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialConversion.R
\name{ConvertH5ADSpatialToSeurat}
\alias{ConvertH5ADSpatialToSeurat}
\title{Convert spatial coordinates from h5ad to Seurat format}
\usage{
ConvertH5ADSpatialToSeurat(
  h5ad_file,
  seurat_obj = NULL,
  assay_name = "Spatial",
  verbose = TRUE
)
}
\arguments{
\item{h5ad_file}{H5AD file handle or path}

\item{seurat_obj}{Seurat object to add spatial data to}

\item{assay_name}{Name of the assay to associate spatial data with}

\item{verbose}{Print progress messages}
}
\value{
Modified Seurat object with spatial data
}
\description{
Convert spatial coordinates from h5ad to Seurat format
}
