% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5MU.R
\name{ConvertH5MUSpatialToSeurat}
\alias{ConvertH5MUSpatialToSeurat}
\title{Convert h5mu spatial data to Seurat format for a specific modality}
\usage{
ConvertH5MUSpatialToSeurat(
  h5mu_file,
  seurat_obj,
  modality,
  assay_name,
  verbose = FALSE
)
}
\arguments{
\item{h5mu_file}{H5File connection to h5mu file}

\item{seurat_obj}{Seurat object}

\item{modality}{Modality name in h5mu file}

\item{assay_name}{Corresponding assay name in Seurat}

\item{verbose}{Logical}
}
\value{
Modified Seurat object
}
\description{
Convert h5mu spatial data to Seurat format for a specific modality
}
\keyword{internal}
