% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialConversion.R
\name{ConvertSeuratSpatialToH5AD}
\alias{ConvertSeuratSpatialToH5AD}
\title{Convert Seurat spatial data to h5ad format}
\usage{
ConvertSeuratSpatialToH5AD(
  seurat_obj,
  h5ad_file,
  library_id = "library_1",
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object with spatial data}

\item{h5ad_file}{H5AD file handle or path to write to}

\item{library_id}{Library ID for spatial data (default: "library_1")}

\item{verbose}{Print progress messages}
}
\description{
Convert Seurat spatial data to h5ad format
}
