% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsH5Access.R
\name{CopyH5MatrixData}
\alias{CopyH5MatrixData}
\title{Copy HDF5 matrix data handling nested paths}
\usage{
CopyH5MatrixData(src_group, src_path, dst_loc, dst_name, verbose = FALSE)
}
\arguments{
\item{src_group}{The source H5Group containing the data}

\item{src_path}{Path to the source data (can be nested)}

\item{dst_loc}{Destination H5File or H5Group}

\item{dst_name}{Name for the copied data at destination}

\item{verbose}{Show progress messages}
}
\value{
Invisible NULL
}
\description{
Copy HDF5 matrix data handling nested paths
}
\keyword{internal}
