% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSpatialDemoData.R
\name{create_synthetic_spatial}
\alias{create_synthetic_spatial}
\title{Create Synthetic Spatial Data for Testing}
\usage{
create_synthetic_spatial(
  n_spots = 100,
  n_genes = 200,
  technology = c("Visium", "SlideSeq")
)
}
\arguments{
\item{n_spots}{Number of spatial spots}

\item{n_genes}{Number of genes}

\item{technology}{Type of spatial technology to simulate}
}
\value{
Seurat object with simulated spatial data
}
\description{
Creates a minimal Seurat object with spatial information for testing
}
