% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5Seurat.R
\name{DetectMultiLibrary}
\alias{DetectMultiLibrary}
\title{Detect if h5Seurat file contains multi-library spatial data}
\usage{
DetectMultiLibrary(file)
}
\arguments{
\item{file}{h5Seurat file object}
}
\value{
List with is_multilibrary and library_col
}
\description{
Detect if h5Seurat file contains multi-library spatial data
}
\keyword{internal}
