% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialConversion.R
\name{DetectSpatialTechnology}
\alias{DetectSpatialTechnology}
\title{Detect spatial technology from data structure}
\usage{
DetectSpatialTechnology(coords, metadata = NULL)
}
\arguments{
\item{coords}{Spatial coordinates matrix}

\item{metadata}{Additional metadata}
}
\value{
Technology type string
}
\description{
Detect spatial technology from data structure
}
\keyword{internal}
