% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFOV.R
\name{ExtractFOVCentroids}
\alias{ExtractFOVCentroids}
\title{Functions for handling FOV (Field of View) spatial objects}
\usage{
ExtractFOVCentroids(fov_obj)
}
\arguments{
\item{fov_obj}{FOV object from Seurat}
}
\value{
Matrix of centroids (cells x 2) with x,y coordinates
}
\description{
FOV objects are used for imaging-based spatial transcriptomics
(Xenium, MERFISH, CosMx, etc.)
}
\details{
Extract centroids from FOV object
}
\keyword{internal}
