% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFOV.R
\name{ExtractFOVMolecules}
\alias{ExtractFOVMolecules}
\title{Extract molecule positions from FOV object}
\usage{
ExtractFOVMolecules(fov_obj)
}
\arguments{
\item{fov_obj}{FOV object from Seurat}
}
\value{
Data frame with molecule positions or NULL
}
\description{
Extract molecule positions from FOV object
}
\keyword{internal}
