% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsAssayCompat.R
\name{GetAssayDataCompat}
\alias{GetAssayDataCompat}
\title{Get assay data with V4/V5 compatibility}
\usage{
GetAssayDataCompat(object, layer_or_slot, verbose = FALSE)
}
\arguments{
\item{object}{Assay or Assay5 object}

\item{layer_or_slot}{Name of layer (V5) or slot (V4) to retrieve}

\item{verbose}{Print diagnostic messages}
}
\value{
Matrix or sparse matrix of assay data
}
\description{
Retrieves data from an Assay or Assay5 object, automatically using the
appropriate parameter name (layer for V5, slot for V4).
}
\seealso{
\code{\link{SafeGetAssayData}} in V5Compatibility.R for the
  original implementation
}
\keyword{internal}
