% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsAssayCompat.R
\name{GetAssayLayersCompat}
\alias{GetAssayLayersCompat}
\title{Get available layers or slots from assay}
\usage{
GetAssayLayersCompat(object, verbose = FALSE)
}
\arguments{
\item{object}{Assay or Assay5 object}

\item{verbose}{Print diagnostic messages}
}
\value{
Character vector of layer/slot names
}
\description{
Returns list of available data layers (V5) or slots (V4) from an Assay
or Assay5 object.
}
\keyword{internal}
