% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GetCompressionLevel}
\alias{GetCompressionLevel}
\title{Get compression level for HDF5 dataset creation}
\usage{
GetCompressionLevel()
}
\value{
Integer gzip compression level (0-9)
}
\description{
Get compression level for HDF5 dataset creation
}
\keyword{internal}
