% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveH5MU.R
\name{GetDefaultAssayToModalityMapping}
\alias{GetDefaultAssayToModalityMapping}
\title{Get default assay to modality name mapping}
\usage{
GetDefaultAssayToModalityMapping(assays)
}
\arguments{
\item{assays}{Character vector of assay names}
}
\value{
Named character vector mapping assay names to modality names
}
\description{
Get default assay to modality name mapping
}
\keyword{internal}
