% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5MU.R
\name{GetDefaultModalityMapping}
\alias{GetDefaultModalityMapping}
\title{Get default modality to assay name mapping}
\usage{
GetDefaultModalityMapping(modalities)
}
\arguments{
\item{modalities}{Character vector of modality names}
}
\value{
Named character vector mapping modality names to assay names
}
\description{
Get default modality to assay name mapping
}
\keyword{internal}
