% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5LayerSupport.R
\name{GetLayerPath}
\alias{GetLayerPath}
\title{Get Layer Path}
\usage{
GetLayerPath(h5_group, slot_name, verbose = FALSE)
}
\arguments{
\item{h5_group}{The HDF5 assay group}

\item{slot_name}{Name of the slot/layer}

\item{verbose}{Show progress messages}
}
\value{
The path to the data, or NULL if not found
}
\description{
Get the correct path for a layer/slot in V4 vs V5 format
}
\keyword{internal}
