% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialConversion.R
\name{GetScaleFactors}
\alias{GetScaleFactors}
\title{Get scale factors from Visium object}
\usage{
GetScaleFactors(visium_obj)
}
\arguments{
\item{visium_obj}{Visium image object}
}
\value{
List of scale factors
}
\description{
Get scale factors from Visium object
}
\keyword{internal}
