% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5Compatibility.R
\name{GetSeuratSlotMapping}
\alias{GetSeuratSlotMapping}
\title{Map all Seurat object slots to h5Seurat/h5ad}
\usage{
GetSeuratSlotMapping(format = c("h5seurat", "h5ad"))
}
\arguments{
\item{format}{Target format ("h5seurat" or "h5ad")}
}
\value{
A list describing the slot mapping
}
\description{
Map all Seurat object slots to h5Seurat/h5ad
}
