% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{GetStringType}
\alias{GetStringType}
\title{Check if an HDF5 dataset is a factor}
\usage{
GetStringType(stype)
}
\arguments{
\item{stype}{An HDF5 dataset or type}
}
\value{
String representation of the type
}
\description{
Get string type from HDF5 dataset
}
\keyword{internal}
