% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFOV.R
\name{H5ADToFOV}
\alias{H5ADToFOV}
\title{Convert h5ad spatial data to FOV object}
\usage{
H5ADToFOV(spatial_data, key = "fov")
}
\arguments{
\item{spatial_data}{List with spatial coordinates and optionally segmentation}

\item{key}{FOV key/name}
}
\value{
FOV object or NULL
}
\description{
Convert h5ad spatial data to FOV object
}
\note{
This is a placeholder - full FOV reconstruction requires SeuratObject functions
}
\keyword{internal}
