% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{H5MUToH5Seurat}
\alias{H5MUToH5Seurat}
\title{Convert H5MU files to h5Seurat files}
\usage{
H5MUToH5Seurat(source, dest, assay = "RNA", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source}{Source dataset: a Seurat object, filename path, or H5File connection}

\item{dest}{Name/path of destination file. If only a file type is provided
(e.g., "h5seurat", "h5ad", "loom"), the extension is appended to the source
filename (for file sources) or the Seurat project name (for Seurat objects).
Supported formats: h5seurat, h5ad, loom}

\item{assay}{For h5Seurat -> other formats: name of assay to convert.
For other formats -> h5Seurat: name to assign to the assay.
Default is "RNA".}

\item{overwrite}{Logical; if \code{TRUE}, overwrite an existing destination file.
Default is \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE} (default), show progress updates}
}
\value{
Returns a handle to \code{dest} as an \code{\link{h5Seurat}} object
}
\description{
Convert H5MU files to h5Seurat files
}
\keyword{internal}
