% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Seurat_bindings.R
\name{h5Seurat-bindings}
\alias{h5Seurat-bindings}
\alias{Cells.h5Seurat}
\alias{Cells}
\alias{DefaultAssay.h5Seurat}
\alias{DefaultAssay<-.h5Seurat}
\alias{Idents.h5Seurat}
\alias{Idents}
\alias{IsGlobal.H5Group}
\alias{IsGlobal}
\alias{Key.H5Group}
\alias{Key}
\alias{Project.h5Seurat}
\alias{Project}
\alias{Project<-.h5Seurat}
\alias{Stdev.h5Seurat}
\title{Seurat bindings for h5Seurat files}
\usage{
\method{Cells}{h5Seurat}(x, ...)

\method{DefaultAssay}{h5Seurat}(object, ...)

\method{DefaultAssay}{h5Seurat}(object, ...) <- value

\method{Idents}{h5Seurat}(object, ...)

\method{IsGlobal}{H5Group}(object, ...)

\method{Key}{H5Group}(object, ...)

\method{Project}{h5Seurat}(object, ...)

\method{Project}{h5Seurat}(object, ...) <- value

\method{Stdev}{h5Seurat}(object, reduction = "pca", ...)
}
\arguments{
\item{x}{An h5Seurat object}

\item{object}{An h5Seurat or H5Group object}

\item{reduction}{Name of the dimensional reduction (default: "pca")}

\item{value}{Value to set}

\item{...}{Additional arguments passed to methods}
}
\description{
Seurat bindings for h5Seurat files
}
