% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsAssayCompat.R
\name{IsAssay5}
\alias{IsAssay5}
\title{Check if object is Assay5 (V5 format)}
\usage{
IsAssay5(object)
}
\arguments{
\item{object}{Object to check}
}
\value{
Logical indicating if object inherits from Assay5
}
\description{
Check if object is Assay5 (V5 format)
}
\keyword{internal}
