% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{IsCompound}
\alias{IsCompound}
\title{Check for compound type in HDF5 dataset}
\usage{
IsCompound(x)
}
\arguments{
\item{x}{An HDF5 dataset}
}
\value{
TRUE if the dataset has a compound type
}
\description{
Check for compound type in HDF5 dataset
}
\keyword{internal}
