% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5DAccess.R
\name{IsEffectively1D}
\alias{IsEffectively1D}
\title{Check if H5D dataset is effectively 1D}
\usage{
IsEffectively1D(dataset)
}
\arguments{
\item{dataset}{An H5D dataset object}
}
\value{
Logical indicating if dataset is effectively 1D
}
\description{
Check if H5D dataset is effectively 1D
}
\keyword{internal}
