% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{IsFactor}
\alias{IsFactor}
\title{Check if an HDF5 dataset is a factor}
\usage{
IsFactor(x)
}
\arguments{
\item{x}{An HDF5 dataset}
}
\value{
TRUE if the dataset represents a factor
}
\description{
Check if an HDF5 dataset is a factor
}
\keyword{internal}
