% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsMatrix.R
\name{IsMatrixEmptyUtil}
\alias{IsMatrixEmptyUtil}
\title{Check if matrix is empty or all zeros}
\usage{
IsMatrixEmptyUtil(mat)
}
\arguments{
\item{mat}{Matrix to check}
}
\value{
Logical indicating if matrix is empty
}
\description{
Check if matrix is empty or all zeros
}
\keyword{internal}
