% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5LayerSupport.R
\name{IsV5Structure}
\alias{IsV5Structure}
\title{Check if h5Seurat file uses V5 structure}
\usage{
IsV5Structure(h5_file)
}
\arguments{
\item{h5_file}{Path to h5Seurat file or h5Seurat object}
}
\value{
Logical indicating if file uses V5 structure
}
\description{
Check if h5Seurat file uses V5 structure
}
\keyword{internal}
