% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatRegistry.R
\name{ListFormats}
\alias{ListFormats}
\title{List all registered format conversions}
\usage{
ListFormats()
}
\value{
A list with loaders, savers, and direct_paths names
}
\description{
List all registered format conversions
}
\keyword{internal}
