% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_bindings.R
\name{loom-bindings}
\alias{loom-bindings}
\alias{DefaultAssay.loom}
\alias{dim.loom}
\alias{dim}
\title{Seurat binding for loom files}
\usage{
\method{DefaultAssay}{loom}(object, ...)

\method{dim}{loom}(x)
}
\arguments{
\item{object}{A connected \code{loom} file handle}

\item{...}{Additional arguments passed to Seurat generics}

\item{x}{A \code{loom} object}
}
\description{
Seurat binding for loom files
}
