% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveH5MU.R
\name{PrepareSpatialForH5MU}
\alias{PrepareSpatialForH5MU}
\title{Prepare spatial data for h5mu export}
\usage{
PrepareSpatialForH5MU(object, assays, modality.names, verbose = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{assays}{Assay names}

\item{modality.names}{Modality name mapping}

\item{verbose}{Logical}
}
\value{
Modified Seurat object with spatial metadata prepared
}
\description{
Prepare spatial data for h5mu export
}
\keyword{internal}
