% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5MU.R
\name{ReadGlobalObs}
\alias{ReadGlobalObs}
\title{Read global observation metadata from h5mu file}
\usage{
ReadGlobalObs(h5mu, verbose = FALSE)
}
\arguments{
\item{h5mu}{H5File connection to h5mu file}

\item{verbose}{Logical; print messages}
}
\value{
Data frame with global cell metadata, or NULL
}
\description{
Read global observation metadata from h5mu file
}
\keyword{internal}
