% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{ReadH5}
\alias{ReadH5}
\alias{as.array.H5D}
\alias{as.array}
\alias{as.data.frame.H5D}
\alias{as.data.frame.H5Group}
\alias{as.list.H5D}
\alias{as.list.H5Group}
\alias{as.matrix.H5D}
\alias{as.matrix}
\alias{as.matrix.H5Group}
\title{Read HDF5 Files}
\usage{
\method{as.array}{H5D}(x, ...)

\method{as.data.frame}{H5D}(x, row.names = NULL, optional = FALSE, ...)

\method{as.data.frame}{H5Group}(x, row.names = NULL, optional = FALSE, ...)

\method{as.list}{H5D}(x, ...)

\method{as.list}{H5Group}(x, recursive = TRUE, ...)

\method{as.matrix}{H5D}(x, transpose = FALSE, ...)

\method{as.matrix}{H5Group}(x, ...)
}
\arguments{
\item{x}{An HDF5 dataset (H5D) object}

\item{...}{Arguments passed to other methods}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{transpose}{Transpose the matrix before returning}
}
\value{
Varies depending on the method being called

\code{as.array}: an array with the data from the HDF5 dataset

\code{as.data.frame}: returns a \code{\link[base]{data.frame}} with
the data from the HDF5 dataset

\code{as.list}: a list with the data from the HDF5 dataset

\code{as.matrix}: a matrix with the data from the HDF5 dataset
}
\description{
Read data from HDF5 files
}
\seealso{
\code{\link{as.sparse}}
}
