% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5Compatibility.R
\name{ReadH5Direct}
\alias{ReadH5Direct}
\title{Read HDF5 dataset with V5 compatibility}
\usage{
ReadH5Direct(dataset, verbose = FALSE)
}
\arguments{
\item{dataset}{An H5D dataset object}

\item{verbose}{Logical, print messages}
}
\value{
The data read from the dataset
}
\description{
Read HDF5 dataset with V5 compatibility
}
\keyword{internal}
