% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5Seurat.R
\name{ReadLibraryIds}
\alias{ReadLibraryIds}
\title{Read library IDs from h5Seurat metadata}
\usage{
ReadLibraryIds(file, library_col)
}
\arguments{
\item{file}{h5Seurat file object}

\item{library_col}{Name of the library ID column}
}
\value{
Character vector of library IDs
}
\description{
Read library IDs from h5Seurat metadata
}
\keyword{internal}
