% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssembleObject.R
\name{ReadSparseMatrix}
\alias{ReadSparseMatrix}
\title{Read sparse matrix from HDF5}
\usage{
ReadSparseMatrix(h5_group, verbose = FALSE)
}
\arguments{
\item{h5_group}{The HDF5 group containing sparse matrix data}

\item{verbose}{Show progress updates}
}
\value{
A matrix object, or NULL if reading fails
}
\description{
Internal function to read sparse matrix data from h5Seurat files,
handling both V4 and V5 formats
}
\keyword{internal}
