% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatRegistry.R
\name{RegisterFormat}
\alias{RegisterFormat}
\title{Register a file format with its load and save functions}
\usage{
RegisterFormat(ext, loader = NULL, saver = NULL)
}
\arguments{
\item{ext}{Lowercase file extension (e.g., "h5seurat", "h5ad", "loom", "rds")}

\item{loader}{Function(file, assay, verbose, ...) -> Seurat object}

\item{saver}{Function(object, dest, overwrite, verbose, ...) -> invisible(dest)}
}
\description{
Register a file format with its load and save functions
}
\keyword{internal}
