% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5MU.R
\name{RestoreSpatialFromH5MU}
\alias{RestoreSpatialFromH5MU}
\title{Restore spatial data from h5mu file to Seurat object}
\usage{
RestoreSpatialFromH5MU(
  h5mu_file,
  seurat_obj,
  modalities,
  assay.names,
  verbose = FALSE
)
}
\arguments{
\item{h5mu_file}{H5File connection to h5mu file}

\item{seurat_obj}{Seurat object to add spatial data to}

\item{modalities}{Character vector of modalities that were loaded}

\item{assay.names}{Named vector mapping modality to assay names}

\item{verbose}{Logical; print messages}
}
\value{
Modified Seurat object with spatial data
}
\description{
Restore spatial data from h5mu file to Seurat object
}
\keyword{internal}
