% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5Compatibility.R
\name{SafeGetAssayData}
\alias{SafeGetAssayData}
\title{Safely get assay data using layer or slot parameter}
\usage{
SafeGetAssayData(object, layer)
}
\arguments{
\item{object}{An Assay or Assay5 object}

\item{layer}{Layer name (preferred for V5)}
}
\value{
Matrix data or NULL
}
\description{
Safely get assay data using layer or slot parameter
}
\keyword{internal}
