% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5Compatibility.R
\name{SafeGetLayers}
\alias{SafeGetLayers}
\title{Get layers from an Assay5 object safely}
\usage{
SafeGetLayers(object)
}
\arguments{
\item{object}{An Assay or Assay5 object}
}
\value{
Character vector of available layers
}
\description{
Get layers from an Assay5 object safely
}
\keyword{internal}
