% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5DAccess.R
\name{SafeH5DRead}
\alias{SafeH5DRead}
\title{Safely read an H5D dataset}
\usage{
SafeH5DRead(dataset, subset = NULL)
}
\arguments{
\item{dataset}{An H5D dataset object}

\item{subset}{Optional subset indices}
}
\value{
The data from the dataset
}
\description{
Safely read an H5D dataset
}
\keyword{internal}
