% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{SafeSetLayerData}
\alias{SafeSetLayerData}
\title{Safely set layer data in a V5 Assay}
\usage{
SafeSetLayerData(object, layer, value)
}
\arguments{
\item{object}{An Assay or Assay5 object}

\item{layer}{Layer name}

\item{value}{Data to set}
}
\value{
Modified assay object
}
\description{
Handles the case where LayerData<- might not be exported from SeuratObject
}
\keyword{internal}
