% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialConversion.R
\name{SanitizeImageKey}
\alias{SanitizeImageKey}
\title{Sanitize string to valid Seurat image key}
\usage{
SanitizeImageKey(x)
}
\arguments{
\item{x}{String to sanitize}
}
\value{
Valid Seurat key ending with underscore
}
\description{
Sanitize string to valid Seurat image key
}
\keyword{internal}
