% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsAssayCompat.R
\name{SetAssayDataCompat}
\alias{SetAssayDataCompat}
\title{Set assay data with V4/V5 compatibility}
\usage{
SetAssayDataCompat(object, layer_or_slot, new.data, verbose = FALSE)
}
\arguments{
\item{object}{Assay or Assay5 object}

\item{layer_or_slot}{Name of layer (V5) or slot (V4) to set}

\item{new.data}{New data to set}

\item{verbose}{Print diagnostic messages}
}
\value{
Modified Assay or Assay5 object
}
\description{
Sets data in an Assay or Assay5 object, automatically using the appropriate
parameter name (layer for V5, slot for V4).
}
\keyword{internal}
