% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{srtdisk-package}
\alias{srtdisk-package}
\alias{srtdisk}
\alias{SeuratDisk}
\title{srtdisk: Extended HDF5-Based Single Cell File Format Support}
\description{
Extended implementation of SeuratDisk for Seurat v5. Provides interfaces for
HDF5-based single-cell file formats including h5Seurat and AnnData (h5ad)
with enhanced support for Seurat v5 Assay5 and spatial data.
}
\section{Package options}{


srtdisk uses the following options to control behavior, users can configure
these with \code{\link[base]{options}}:

\describe{
 \item{\code{SeuratDisk.dtypes.logical_to_int}}{
  When writing \link[base]{logical} vectors, coerce to integer types to
  ensure compatibility across languages (see \code{\link{BoolToInt}} for
  more details)
 }
 \item{\code{SeuratDisk.dtypes.dataframe_as_group}}{
  When writing \link[base]{data.frame}s, always write out as a group
  regardless of factor presence
 }
 \item{\code{SeuratDisk.chunking.MARGIN}}{
  Default direction for chunking datasets; choose from:
  \describe{
   \item{largest}{Chunk along the largest dimension of a dataset}
   \item{smallest}{Chunk along the smallest dimension}
   \item{first}{Chunk along the first dimension}
   \item{last}{Chunk along the last dimension}
  }
 }
 \item{\code{SeuratDisk.dimreducs.allglobal}}{
  Treat all DimReducs as global, regardless of actual global status
 }
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mianaz.github.io/srtdisk/}
  \item \url{https://github.com/mianaz/srtdisk}
  \item Report bugs at \url{https://github.com/mianaz/srtdisk/issues}
}

}
\author{
\strong{Maintainer}: Miana Z \email{zzeng4@nd.edu}

Authors:
\itemize{
  \item Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID})
}

}
