% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestObject.R
\name{TestObject}
\alias{TestObject}
\alias{IsDataFrame}
\alias{IsFactor}
\alias{IsList}
\alias{IsLogical}
\alias{IsMatrix}
\title{Test an object's class}
\usage{
IsDataFrame()

IsFactor()

IsList()

IsLogical()

IsMatrix()
}
\description{
Generic functions for testing if an object satisfies class membership
}
\keyword{internal}
