% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5Compatibility.R
\name{TransferMetadataV5}
\alias{TransferMetadataV5}
\title{Transfer metadata from V5 h5Seurat to h5ad}
\usage{
TransferMetadataV5(source, dfile, dname = "obs", index = NULL, verbose = TRUE)
}
\arguments{
\item{source}{Source H5 group (potentially wrapped in environment)}

\item{dfile}{Destination H5 file}

\item{dname}{Destination group name}

\item{index}{Cell index for subsetting}

\item{verbose}{Logical, print messages}
}
\value{
NULL invisibly
}
\description{
Transfer metadata from V5 h5Seurat to h5ad
}
\keyword{internal}
