% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5MU.R
\name{TryAddVisiumImagesFromH5MU}
\alias{TryAddVisiumImagesFromH5MU}
\title{Try to add Visium images from h5mu structure}
\usage{
TryAddVisiumImagesFromH5MU(
  seurat_obj,
  spatial_uns,
  coords,
  visium_data,
  assay_name,
  verbose = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{spatial_uns}{HDF5 group containing spatial/uns data}

\item{coords}{Spatial coordinates matrix}

\item{visium_data}{Visium metadata list}

\item{assay_name}{Assay name}

\item{verbose}{Logical}
}
\value{
Modified Seurat object
}
\description{
Try to add Visium images from h5mu structure
}
\keyword{internal}
