% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{ValiateLoom}
\alias{ValiateLoom}
\alias{LoomValidate0.1}
\alias{ValidateLoom}
\alias{LoomValidate3.0.0}
\title{Validate Loom Files}
\usage{
LoomValidate0.1(lfile, verbose = TRUE)

LoomValidate3.0.0(lfile, verbose = TRUE)
}
\arguments{
\item{lfile}{A loom file object or path}

\item{verbose}{Whether to print validation messages}
}
\value{
Invisibly returns TRUE if valid, throws an error if invalid
}
\description{
Functions for validating loom files before connection
}
\seealso{
\href{http://linnarssonlab.org/loompy/format/index.html}{Loom file format}
}
\keyword{internal}
