% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveH5MU.R
\name{ValidateMultimodalObject}
\alias{ValidateMultimodalObject}
\title{Validate Seurat object for multimodal export}
\usage{
ValidateMultimodalObject(object, assays, verbose = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{assays}{Character vector of assays to validate}

\item{verbose}{Logical}
}
\value{
List with 'valid' (logical) and 'message' (character) elements
}
\description{
Validate Seurat object for multimodal export
}
\keyword{internal}
