% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{WriteAttribute}
\alias{WriteAttribute}
\title{Write an attribute to an HDF5 file, group, or dataset}
\usage{
WriteAttribute(x, name, lfile, stype, transpose = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{An object to write out}

\item{name}{Name to store attribute as}

\item{lfile}{An HDF5 \link[hdf5r:H5File]{file}, \link[hdf5r:H5Group]{group}, or
\link[hdf5r:H5D]{dataset}}

\item{stype}{Data type of attribute}

\item{transpose}{Whether to transpose the data}

\item{verbose}{Whether to print messages}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Write an attribute to an HDF5 file, group, or dataset
}
\keyword{internal}
