% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteH5Group.R
\name{WriteH5GroupAssay5}
\alias{WriteH5GroupAssay5}
\title{Handle V5-specific assay writing}
\usage{
WriteH5GroupAssay5(x, name, hgroup, verbose = TRUE)

WriteH5GroupAssay5(x, name, hgroup, verbose = TRUE)
}
\arguments{
\item{x}{An object}

\item{name}{Name to save data as}

\item{hgroup}{An HDF5 file or group (\code{H5File} or \code{H5Group} objects
from hdf5r)}

\item{verbose}{Show progress updates}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Handle V5-specific assay writing

Write V5 Assay objects to HDF5
}
\keyword{internal}
