% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5Compatibility.R
\name{WriteH5SpatialV5}
\alias{WriteH5SpatialV5}
\title{Handle spatial data for V5 objects}
\usage{
WriteH5SpatialV5(object, hgroup, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{hgroup}{An H5Group to write to}

\item{verbose}{Logical, print messages}
}
\value{
NULL invisibly
}
\description{
Handle spatial data for V5 objects
}
\keyword{internal}
