% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5LayerSupport.R
\name{WriteV5Layer}
\alias{WriteV5Layer}
\title{Write V5 Layer Data}
\usage{
WriteV5Layer(
  h5_group,
  layer_name,
  matrix_data,
  features = NULL,
  verbose = FALSE
)
}
\arguments{
\item{h5_group}{The HDF5 group to write to}

\item{layer_name}{Name of the layer (e.g., "counts", "data", "scale.data")}

\item{matrix_data}{The matrix data to write}

\item{features}{Character vector of feature names}

\item{verbose}{Show progress messages}
}
\value{
Invisible NULL
}
\description{
Write matrix data in V5 layer format (sparse or dense)
}
\keyword{internal}
