% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendData.R
\name{AppendData}
\alias{AppendData}
\alias{AppendData.character}
\alias{AppendData.H5File}
\alias{AppendData.h5Seurat}
\title{Append data from an h5Seurat file to a preexisting
\code{\link[Seurat]{Seurat}} object}
\usage{
AppendData(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{AppendData}{character}(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{AppendData}{H5File}(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{AppendData}{h5Seurat}(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file}{Name of an h5Seurat file path (character) or connected h5Seurat file}

\item{object}{A \code{\link[Seurat]{Seurat}} object to append data to}

\item{assays}{One of:
\itemize{
  \item A character vector with names of assays
  \item A character vector with one or more of \code{counts}, \code{data},
  \code{scale.data} describing which slots of \strong{all assays} to load
  \item A named list where each entry is either the name of an assay or a vector
  describing which slots (described above) to take from which assay
  \item \code{NULL} for all assays
  \item \code{FALSE} for no assays
}}

\item{reductions}{Which dimensional reductions to load; pass \code{NULL} for all,
a character vector for specific reductions, or \code{FALSE} for no reductions}

\item{graphs}{Which graphs to load; pass \code{NULL} for all,
a character vector for specific graphs, or \code{FALSE} for no graphs}

\item{images}{Which images to load; pass \code{NULL} for all,
a character vector for specific images, or \code{FALSE} for no images}

\item{extras}{Extra information to load; supports any combination of the
following values:
\describe{
 \item{\dQuote{commands}}{Load command logs. If \code{overwrite = TRUE},
 replaces existing command logs}
}}

\item{overwrite}{Overwrite existing data in \code{object} with data from
\code{file}}

\item{verbose}{Display progress messages}

\item{...}{Additional arguments passed to methods}
}
\value{
\code{object} with the extra data requested
}
\description{
Append data from an h5Seurat file to a preexisting
\code{\link[Seurat]{Seurat}} object
}
