% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R, R/ReadH5.R
\name{CompoundToGroup}
\alias{CompoundToGroup}
\title{Convert an HDF5 compound dataset to a group}
\usage{
CompoundToGroup(src, dst, dname, order, index = NULL)

CompoundToGroup(src, dst, dname, order, index = NULL)
}
\arguments{
\item{src}{Source compound dataset}

\item{dst}{Destination file}

\item{dname}{Destination name}

\item{order}{Order attribute name}

\item{index}{Index for subsetting}

\item{dest}{An HDF5 file (\code{\link[hdf5r]{H5File}}) or group
(\code{\link[hdf5r]{H5Group}})}

\item{dst.name}{Name of group in \code{dest}}

\item{overwrite}{Overwrite existing group \code{dst.name} in \code{dest}}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Convert an HDF5 compound dataset to a group

Convert compound dataset to group
}
\keyword{internal}
