% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{CompoundToGroup}
\alias{CompoundToGroup}
\title{Convert an HDF5 compound dataset to a group}
\usage{
CompoundToGroup(src, dst, dname, order, index = NULL, name_map_fn = NULL)
}
\arguments{
\item{src}{An HDF5 dataset (\code{\link[hdf5r]{H5D}}) of type
\code{\link[hdf5r]{H5T_COMPOUND}}}

\item{dst}{An HDF5 file (\code{\link[hdf5r]{H5File}}) or group
(\code{\link[hdf5r]{H5Group}})}

\item{dname}{Name of group in \code{dst}}

\item{order}{Column order specification}

\item{index}{Integer values of which values to pull; defaults to all values}

\item{name_map_fn}{Optional function to map compound field names}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Convert an HDF5 compound dataset to a group
}
\keyword{internal}
