% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{H5MUToH5AD}
\alias{H5MUToH5AD}
\title{Convert H5MU files to H5AD files (extract single modality)}
\usage{
H5MUToH5AD(source, dest, modality = "rna", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source}{Source dataset: a Seurat object, filename path, or H5File connection}

\item{dest}{Name/path of destination file. If only a file type is provided
(e.g., "h5seurat", "h5ad", "loom"), the extension is appended to the source
filename (for file sources) or the Seurat project name (for Seurat objects).
Supported formats: h5seurat, h5ad, loom}

\item{modality}{Name of modality to extract from h5mu file}

\item{overwrite}{Logical; if \code{TRUE}, overwrite an existing destination file.
Default is \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE} (default), show progress updates}
}
\value{
Returns a handle to \code{dest} as an \code{\link[hdf5r]{H5File}} object
}
\description{
Convert H5MU files to H5AD files (extract single modality)
}
\keyword{internal}
