% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{H5MUToH5AD}
\alias{H5MUToH5AD}
\title{Convert H5MU files to H5AD files (extract single modality)}
\usage{
H5MUToH5AD(source, dest, modality = "rna", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source}{Source dataset}

\item{dest}{Name of destination dataset}

\item{modality}{Name of modality to extract from h5mu file}

\item{overwrite}{Overwrite existing \code{dest}}

\item{verbose}{Show progress updates}
}
\value{
Returns a handle to \code{dest} as an \code{\link[hdf5r]{H5File}} object
}
\description{
Convert H5MU files to H5AD files (extract single modality)
}
\keyword{internal}
