% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{H5MUToH5Seurat}
\alias{H5MUToH5Seurat}
\title{Convert H5MU files to h5Seurat files}
\usage{
H5MUToH5Seurat(source, dest, assay = "RNA", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source}{Source dataset}

\item{dest}{Name of destination dataset}

\item{assay}{Converting from \code{\link{h5Seurat}}: name of assay to write
out; converting to \code{\link{h5Seurat}}: name to store assay data as}

\item{overwrite}{Overwrite existing \code{dest}}

\item{verbose}{Show progress updates}
}
\value{
Returns a handle to \code{dest} as an \code{\link{h5Seurat}} object
}
\description{
Convert H5MU files to h5Seurat files
}
\keyword{internal}
