% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{H5SeuratToH5MU}
\alias{H5SeuratToH5MU}
\title{Convert h5Seurat files to H5MU files}
\usage{
H5SeuratToH5MU(
  source,
  dest,
  assay = DefaultAssay(object = source),
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{source}{Source dataset}

\item{dest}{Name of destination dataset}

\item{assay}{Converting from \code{\link{h5Seurat}}: name of assay to write
out; converting to \code{\link{h5Seurat}}: name to store assay data as}

\item{overwrite}{Overwrite existing \code{dest}}

\item{verbose}{Show progress updates}
}
\value{
Returns a handle to \code{dest} as an \code{\link[hdf5r]{H5File}} object
}
\description{
Convert h5Seurat files to H5MU files
}
\keyword{internal}
