% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5AD.R
\name{LoadH5AD}
\alias{LoadH5AD}
\title{Load an AnnData H5AD file as a Seurat object}
\usage{
LoadH5AD(file, assay.name = "RNA", verbose = TRUE)
}
\arguments{
\item{file}{Path to H5AD file}

\item{assay.name}{Name for the primary assay (default: "RNA")}

\item{verbose}{Show progress messages}
}
\value{
A \code{Seurat} object
}
\description{
Direct conversion from H5AD format to Seurat object without intermediate h5Seurat
}
