% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V5Compatibility.R
\name{WriteH5GroupAssay5}
\alias{WriteH5GroupAssay5}
\title{Write an Assay5 object to an HDF5 group}
\usage{
WriteH5GroupAssay5(x, name, hgroup, verbose = TRUE)
}
\arguments{
\item{x}{An object}

\item{name}{Name to save data as}

\item{hgroup}{An HDF5 file or group (\code{H5File} or \code{H5Group} objects
from hdf5r)}

\item{verbose}{Show progress updates}
}
\value{
Invisibly returns NULL
}
\description{
Write an Assay5 object to an HDF5 group
}
\keyword{internal}
